requirements.txt
streamlit
google-cloud-aiplatform
pandas
sqlalchemy
pytest
python-dotenv
google-generativeai